\name{CastCharacter}
\alias{CastCharacter}
\title{A function to cast factors to strings in a data frame}
\usage{
CastCharacter(d)
}
\arguments{
  \item{d}{data frame}
}
\value{
a data frame with columns as character
}
\description{
A function to cast factors to strings in a data frame
}
\keyword{utilities}

